+++ src/router/Routes.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import CategoryScreen from "../screens/CategoryScreen";
 import MainScreen from "../screens/MainScreen";
-import { routeURL } from "./routePath";
+import { routeURL } from "./paths";
 
 const Routes = () => {
   return (
